+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
+- Trip mines, land mines, and tesla coils now work with non-player entities
+
 ### Deprecated
 
 ### Removed